Minor bugfix/cleanup (#941)
* Fixes slime spawn handling variances

After testing, I was able to see 2 separate problems. 
1: event.getResult() was occasionally returning ALLOW for some of the world spawned slimes. Slime spawn denial is done regardless of the ALLOW/DENY flag. This does not impact slimes spawned via eggs or powered spawners or slimes spawning after a slime is killed which apparently return DEFAULT for getResult().
2: Slimes don't seem to properly handle a getResult() value of DENY. The checkSpawn event is not cancellable. To fix this, the entity is also destroyed by setting isDead to true. This does occasionally cause slimes to appear on the minimap briefly before disappearing.

Lastly, with a forge update this could be handled much better by using the EntitySlime.spawnReason enum. That is not present in the current forge version, so I would consider this a temporary workaround and this method could/should be reworked if/when forge is updated.

* Make repellators stop slimes.

While testing, this was used to name the slimes as they were being generated. It turns out to also make the slimes behave properly in addition to the other changes.

* Minor Cleanup

Moved the slime handling out of the rest of the method and takes an approach of setting it to DEFAULT of it is ALLOW and then also adding a custom name tag if it doesn't have one. If it does have a custom name tag and ALLOW, then the slime spawn will be allowed.

* Moved the setCustomNameTag method

This method is still required to properly stop ALL slime spawns but it should only alter the customNameTag when a slime is intended to be stopped from spawning.

* Fixed likely exception

Check for instance of EntitySlime before casting it. This method is only required to stop slimes from spawning so is only being called against slimes.

Co-authored-by: Randy Braunm <randy@unseencolor.com>