Neutronium Fusion Buff (#945)
* Neutronium Fusion Buff

With the recent addition of the Naquadah line and making Naquadria a lot more difficult, it seems fair to buff this recipe for neutronium to not be completely garbage. Americium & Naquadria amount increased from 96L to 144L, and increased Neutronium gain from 24L to 144L.
Time required to craft this is now 24 seconds and requires 122880 EU/t, or a total of 58,982,400 EU (58M) to create one ingot. For reference, it takes 3.2B EU total for a single neutronium hot ingot to smelt with Radon gas in a regular EBF with no overclocks. However, it's also possible that MEBFs or potentially Volcanus's will inevitably out pace this after a certain point with large EU/t gain.

Thoughts on this?

* Retier and large buff

Neutronium now takes 4 seconds, but requires UHV voltage input and 1B EU startup; T4 fusion at the minimum.

* UV input, not UHV, whoopies

* Molten Neutronium is at T3 fusion, again

Requires 640M EU start, ZPM voltage input, and 12 seconds of total processing time for a single ingot. The "difficulty" or challenge of this recipe is the molten Naquadria being required.