feat(vent_cover): cool hot-coolant
Provides a low tech method to cool down hot-coolant using vent covers.
It is meant for reprocessing hot coolant before having access to
a Large Heat Exchanger or a VacuumFreezer.
The tier of vent cover and the type of aire or water in front of it
determines the cooling speed.

Feature preview: https://youtu.be/g5a-WuVBqOI