address ModularUI methods that can only be overridden
A call of super was safely deleted.

A call of draw() is legit, but it's labeled as OverrideOnly in ModularUI,
so IntelliJ considers it incorrect to call draw(). This warning was
suppressed. On top of that, IntelliJ incorrectly says that this
suppression is redundant, so that was addressed too.

Fixing the issue without suppressing the warning entails changing the
ModularUI mod, which is too much work for one warning.
However, if there are many warnings like that, it might be reasonable
to change ModularUI.

The whitespace change is Spotless, although it did not detect it
before the changes to the file.