convert an assert into if
Only tStack 2 is checked for null because if it isn't null then
tStack1 also isn't null based on the "if" above.
Also IntelliJ was sure that tStack is not null for some reason.

On a side note, assertions work only either with a specified flag
or in debug runs. Therefore, it is dangerous to rely on them.