Long Distance Pipeline texture improvements
The side textures of the pipeline start/end blocks are now aligned with
the axis through the input output face (most visible with the Item
Pipeline). This is based on the Buffer texture code.

In addition, the input and output face textures have received the
typical blue/orange coding found on multiblock hatches and busses.