buff rutile yield from garnet processing
in https://github.com/GTNewHorizons/GT5-Unofficial/pull/1927 titaniumtetrachloride was changed to need 3 rutile instead of 1. This was the result of new processing lines for bauxite and ilmenite that yielded much higher amounts. As a result of the changes to titaniumtetrachloride all other sources of rutile were nerfed by a factor of three. This change would return rutile to the same yield of titanium as before the titaniumtetrachloride rebalance