Syncing TextWidget (#34)
* Replaces one instance where a dynamic text field is being communicated
between server and client with a purely client-side method. In some
cases this allows the client to do localization on these texts
correctly. (See https://github.com/GTNewHorizons/ModularUI/pull/69.)

This is a part of the series of PRs implementing new ModularUI features,
however this mod does not use any numeric text fields or similar, and
thus no other changes are needed.