Update GT_MetaTileEntity_IndustrialApiary.java
The Apimancer's Drainer uses GT_MetaTileEntity_IndustrialApiary.getUsedQueen to retrieve the queen currently being processed by the Industrial Apiary.

getUsedQueen is annotated with @SideOnly(Side.CLIENT) which only allows this function to be called from singleplayer. This function is not available when called from server-side and leads to the crash detailed in this issue.

As far as I can tell, this function has no reason to be singleplayer / client-side only. This seems to be an artifact from when this function was imported over 2 years ago.

Testing this change in our world prevents the crash and allows the Apimancer's Drainer to work as expected with the Industrial Apiary.

Resolves GTNewHorizons/GT-New-Horizons-Modpack#16222