Fix crash when configuring ender tank covers
The client was calling DimensionManager.getWorld on the client when it
shouldn't have been.

The fix was to move the private/public toggling code into the cover
update method, so that it gets called on the server instead of the
client. This is slightly less efficient, but the code is trivial so it
doesn't matter.

This change doesn't have any side effects because the relevant code is
only used in the ender tank cover.