Fixed infinite oil rig getting stuck
While I can't reproduce this in a dev environment, oil rigs save their
pumped fluid by its id. If this id shifts, it won't be able to pump
fluid anymore since the underground fluid's id doesn't match the saved
id.

This commit changes oil rigs to hold a reference to a Fluid instead of
an int, and it also removes the saving code entirely. There was no
reason for it to exist since oil rigs will fetch the fluid automatically
when the fluid field is null.