Rotating machines break adjacent enets
When you have a power-accepting machine next to a terminal cable (cable
with only one connection to another cable), rotating the machine will
cause the cable to lose its enet grid. If the machine is connected to
the cable this bug is mitigated.

This causes any machines on the cable to powerfail, since the cable
won't transfer any power.