Deprecate GTLanguageManager
Standard translations with StatCollector have been favored for a long time, so this PR formally deprecates the old system so that new works do not attempt to use it.

Moving away from using it entirely is a large amount of work that we can chip away over time. But ideally this prevents it from growing any more