Add firestone burn effect to GT items
While making it very much harder to work with the stuff,
this at least forms consistency as beforehand, Firestone would "start burning" only after being processed
into a specific gem.

Note: The comb, GT generated ore blocks as well as Raw Ore don't work with this
change as they appear not to be included in the relevant OrePrefixes.
(Should probably be fixed)

Also I would propose that wearing fire hazmat or some other effect should be allowed to prevent
this effect in the long run.

Additionally we might consider a tooltip that explains this, as the "Vanilla" RC items
each have one.