Added Netherite Bee and Prismarine Bee
Per new lines for 2.8, the Prismarine and Netherite Bees are added for players to make the combs to help boost outputs.

These two bees were done per Noc's request.

Thanks to Maya and Noc for helping me figure this out (and helping me to submit my first ever PR that's with actual code in it)