impr(render): reuse SBR contexts instead of creating new instances (#4734)
Co-authored-by: Alexdoru <57050655+Alexdoru@users.noreply.github.com>

This implementation is going to change. Java is currently not equipped to make an efficient use of a persistent object instance like this.
For the time being, this implementation fixes some bugs and improuves rendering in Gregtech ISBRH.
A new implementation more in line with Java philosophie is being worked on.