Fix for item ejecting for high parallels
When a multi has a high number of parallels, it's possible for the item
ejector to never be able to eject one parallel of a recipe. I thought I
fixed this by tweaking the ItemParallelData PriorityQueue sort order,
but it seems that this was still broken in some situations (naqline
centrifuge w/ 72 parallels or batch mode).

To fix this, I changed the priority queue to sort by how many parallels
are remaining per ItemParallelData instead of the absolute number of
items remaining. This should prevent issues when a recipe outputs a
large amount of one output but a small amount of another.