tooltips: clarify ranges covered by different machines
The word 'radius' should really only be applied to spheres and spherical
areas. Conversely the word 'range' can be used for describing the bounds
of a cuboidal area.

This is intended to clear up some confusion about which machines operate
in what kind of area by making the wording (at least somewhat) consistent.

Changes to the overall phrasing were kept minimal however.

Changes:

- Crop Harvester: Radius -> Range, simplify wording a bit
- World Accelerator: Radius -> Range, wording taken from Crop Harvester
- Monster Repellator: Range -> Radius
- Oil Drilling Rig: Radius -> Working Area
- Concrete Backfiller: Radius -> Range ... horizontally