QoL: Screwdrivers on buffer machines
When using buffer-like machines (ie: item buffer, filter, distributor),
you have to configure either an output size parameter on the relative
faces using a screwdriver.
The issue arises when the face you need to configure is not accessible.

Furthermore, the buffer-like machines do not stack when you pick them
up. This is because of their configuration that may diverge. This commit
adds as well a visibility on the configuration to disambiguate why an
instance do not stack.

As I added this configuration on the item tooltip, I've also thought to
add it on WAILA not to rely on modifying the configuration or breaking
the block to check it.